+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   `create sky project` command or via the `Packages > Dart` menu item
 - added a `pub get` and `pub upgrade` context menu off project directories in
   the tree view
+- added code to better recognize when the analysis server terminates
 - added the ability to sort file members (right click in a dart editor and
   choose `Sort Members`)